refactor: about page filename as index
