+++ frontend/src/App.tsx
@@ -19,7 +19,7 @@ import OrgMembersPage from './pages/Orgs/OrgMembersPage'
 import OrgSettingsPage from './pages/Orgs/OrgSettingsPage'
 import StatusPage from './pages/Status'
 import ScrollToTop from './components/ScrollToTop'
-import AboutPage from './pages/About/About'
+import AboutPage from './pages/About'
 import { StaticLayout } from './StaticLayout'
 import NotFoundPage from './pages/NotFound'
 import { PublicRoute } from './components/PublicRoute'
